+++ src/Core/bsd-base64.h
@@ -50,7 +50,7 @@
 
 //#ifndef HAVE___B64_NTOP
 //# ifndef HAVE_B64_NTOP
-int b64_ntop(u_char const *src, size_t srclength, char *target, 
+unsigned long b64_ntop(u_char const *src, size_t srclength, char *target, 
     size_t targsize);
 int b64_pton(char const *src, u_char *target, size_t targsize);
 //# endif /* !HAVE_B64_NTOP */
